requirements.txt
langchain==1.2.10
langchain-chroma==1.1.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.12
langchain-groq==1.1.2
langchain-huggingface==1.2.0
langchain-text-splitters==1.1.0
sentence-transformers==5.2.2
streamlit==1.54.0
huggingface_hub==1.4.1
langchain-huggingface==1.2.0
beautifulsoup4==4.14.3
python-dotenv==1.2.1